pylansforsakringar: Rewrite and update to work in 2017
Update library to work with the new Länsförsäkringar
internet bank. Login and tokens has changed and
there is also a new JSON API used by the browser.
Update the library to use the new tokens and to
use the JSON API to fetch accounts.

Update README and cleanup requirements.